+++ packages/droff-helpers/src/mod.ts
@@ -1,3 +1,4 @@
+export * as Cache from "./cache";
 export * as Flags from "./flags";
 export * as Intents from "./intents";
 export * as Interactions from "./interactions";