+++ packages/droff-interactions/__tests__/droff-interactions.test.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const droffInteractions = require('..');
-
-describe('droff-interactions', () => {
-    it('needs tests');
-});