+++ packages/droff-redis/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/tim-smart/droff/compare/droff-redis@0.5.0...droff-redis@0.5.1) (2022-03-10)
+
+**Note:** Version bump only for package droff-redis
+
+# [0.5.0](https://github.com/tim-smart/droff/compare/droff-redis@0.4.7...droff-redis@0.5.0) (2022-03-10)
+
+**Note:** Version bump only for package droff-redis
+
+## [0.4.7](https://github.com/tim-smart/droff/compare/droff-redis@0.4.6...droff-redis@0.4.7) (2022-03-09)
+
+**Note:** Version bump only for package droff-redis
+
+## [0.4.6](https://github.com/tim-smart/droff/compare/droff-redis@0.4.5...droff-redis@0.4.6) (2022-03-09)
+
+**Note:** Version bump only for package droff-redis
+
 ## [0.4.5](https://github.com/tim-smart/droff/compare/droff-redis@0.4.4...droff-redis@0.4.5) (2022-03-07)
 
 **Note:** Version bump only for package droff-redis